api/requirements.txt
streamlit==1.16.0
pandas
numpy
plotly_express
json5==0.9.6
jsonschema==4.4.0
sqlalchemy
psycopg2
datetime
streamlit_option_menu
openpyxl
xlsxwriter
requests
streamlit-lottie
protobuf
